requirements.txt
Flask
PyMySQL
Flask-SQLAlchemy
